requirements.txt
streamlit==1.34.0
openai==0.27.2
python-dotenv==0.21.0